Set Look and Feel and editor profile via branding / layer.
Set default look and feel via branding property.
Set editor colour profile as attribute of Editor folder in layer.
Remove praxis.live.laf module and move remaining functionality into
praxis.live.core.ui.
Remove unused font colour files.
