+++ branding/modules/org-netbeans-swing-plaf.jar/org/netbeans/swing/plaf/Bundle.properties
@@ -1,0 +1,1 @@
+LookAndFeelClassName=com.formdev.flatlaf.FlatDarkLaf
