+++ nbproject/project.properties
@@ -28,7 +28,6 @@ modules=\
     ${project.org.praxislive.ide.midi}:\
     ${project.org.praxislive.ide.gui}:\
     ${project.org.praxislive.ide.pxr.osc}:\
-    ${project.org.praxislive.ide.laf}:\
     ${project.org.praxislive.ide.model}:\
     ${project.org.praxislive.ide.properties}:\
     ${project.br.pcgl.netbeans}:\
@@ -56,7 +55,6 @@ project.org.praxislive.ide.editor=praxis.live.editor
 project.org.praxislive.ide.editor.saveflash=praxis.live.editor.saveflash
 project.org.praxislive.ide.graph=praxis.live.graph
 project.org.praxislive.ide.gui=praxis.live.gui
-project.org.praxislive.ide.laf=praxis.live.laf
 project.org.praxislive.ide.logging=praxis.live.logging
 project.org.praxislive.ide.midi=praxis.live.midi
 project.org.praxislive.ide.model=praxis.live.model
