+++ praxis.live.core.ui/manifest.mf
@@ -1,5 +1,6 @@
 Manifest-Version: 1.0
 OpenIDE-Module: org.praxislive.ide.core.ui
+OpenIDE-Module-Install: org/praxislive/ide/core/ui/Installer.class
 OpenIDE-Module-Layer: org/praxislive/ide/core/ui/resources/layer.xml
 OpenIDE-Module-Localizing-Bundle: org/praxislive/ide/core/ui/resources/Bundle.properties
 OpenIDE-Module-Specification-Version: 5.0.0
