--- praxis.live.editor/src/org/praxislive/ide/editor/resources/PLTheme-Java-fontsColors.xml
@@ -1,154 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
-
-Copyright 1997-2010 Oracle and/or its affiliates. All rights reserved.
-
-Oracle and Java are registered trademarks of Oracle and/or its affiliates.
-Other names may be trademarks of their respective owners.
-
-
-The contents of this file are subject to the terms of either the GNU
-General Public License Version 2 only ("GPL") or the Common
-Development and Distribution License("CDDL") (collectively, the
-"License"). You may not use this file except in compliance with the
-License. You can obtain a copy of the License at
-http://www.netbeans.org/cddl-gplv2.html
-or nbbuild/licenses/CDDL-GPL-2-CP. See the License for the
-specific language governing permissions and limitations under the
-License.  When distributing the software, include this License Header
-Notice in each file and include the License file at
-nbbuild/licenses/CDDL-GPL-2-CP.  Oracle designates this
-particular file as subject to the "Classpath" exception as provided
-by Oracle in the GPL Version 2 section of the License file that
-accompanied this code. If applicable, add the following below the
-License Header, with the fields enclosed by brackets [] replaced by
-your own identifying information:
-"Portions Copyrighted [year] [name of copyright owner]"
-
-Contributor(s):
-
-The Original Software is NetBeans. The Initial Developer of the Original
-Software is Sun Microsystems, Inc. Portions Copyright 1997-2007 Sun
-Microsystems, Inc. All Rights Reserved.
-
-If you wish your version of this file to be governed by only the CDDL
-or only the GPL Version 2, indicate your decision by adding
-"[Contributor] elects to include this software in this distribution
-under the [CDDL or GPL Version 2] license." If you do not indicate a
-single choice of license, a recipient has the option to distribute
-your version of this file under either the CDDL, the GPL Version 2 or
-to extend the choice of license to its licensees as provided above.
-However, if you add GPL Version 2 code and therefore, elected the GPL
-Version 2 license, then the option applies only if the new code is
-made subject to such option by the copyright holder.
--->
-<!DOCTYPE fontscolors PUBLIC "-//NetBeans//DTD Editor Fonts and Colors settings 1.1//EN" "http://www.netbeans.org/dtds/EditorFontsColors-1_1.dtd">
-
-<fontscolors>
-    <!-- JavaTokenId categories -->
-    <fontcolor name="character" default="char"/>
-    <fontcolor name="errors" default="error"/>
-    <fontcolor name="identifier" default="identifier"/>
-    <fontcolor name="keyword" default="keyword"/>
-    <fontcolor name="keyword-directive" default="keyword"/>
-    <fontcolor name="literal" default="keyword" />
-    <fontcolor name="comment" default="comment"/>
-    <fontcolor name="number" default="number"/>
-    <fontcolor name="operator" default="operator"/>
-    <fontcolor name="string" default="string"/>
-    <fontcolor name="separator" default="separator"/>
-    <fontcolor name="whitespace" default="whitespace"/>
-    
-    <!--"semantic" highlighting:-->
-    <fontcolor name="mod-unused" foreColor="gray"/>
-    <fontcolor name="mod-abstract" />
-    <fontcolor name="mod-field" foreColor="098618" >
-        <font style="bold" />
-    </fontcolor>
-    <fontcolor name="mod-local-variable" />
-    <fontcolor name="mod-parameter" foreColor="A06001"/>
-    <fontcolor name="mod-method" >
-        <font style="bold" />
-    </fontcolor>
-    <fontcolor name="mod-constructor" >
-        <font style="bold" />
-    </fontcolor>
-    <fontcolor name="mod-class" />
-    <fontcolor name="mod-interface" />
-    <fontcolor name="mod-annotation-type" />
-    <fontcolor name="mod-enum" />
-    <fontcolor name="mod-deprecated" strikeThrough="404040" />
-    <fontcolor name="mod-static" >
-        <font style="italic" />
-    </fontcolor>
-    <fontcolor name="mod-private" />
-    <fontcolor name="mod-package-private" />
-    <fontcolor name="mod-protected" />
-    <fontcolor name="mod-public" />
-
-    <!--"semantic" highlighting:-->
-    
-    <fontcolor name="mod-constructor-declaration" default="constructor" />
-    <fontcolor name="mod-constructor" default="constructor" />
-    <fontcolor name="mod-method-declaration" default="method" />
-    <fontcolor name="mod-method" default="method" />
-    <fontcolor name="mod-field-declaration" foreColor="098A18" default="field" />
-    <fontcolor name="mod-field" foreColor="098A18" default="field" />
-    <fontcolor name="mod-local-variable-declaration" default="local-variable" />
-    <fontcolor name="mod-local-variable" default="local-variable" />
-    <fontcolor name="mod-parameter-declaration" foreColor="ec8e01" default="method-parameter" />
-    <fontcolor name="mod-parameter" foreColor="ec8e01" default="method-parameter" />
-
-    <fontcolor name="mod-class-declaration" default="class"/>
-    <fontcolor name="mod-class" default="class" />
-    <fontcolor name="mod-interface-declaration" default="interface"/>
-    <fontcolor name="mod-interface" default="interface" />
-    <fontcolor name="mod-annotation-type-declaration" default="interface"/>
-    <fontcolor name="mod-annotation-type" default="interface" />
-    <fontcolor name="mod-enum-declaration" default="enum"/>
-    <fontcolor name="mod-enum" default="enum" />
-
-    <fontcolor name="mod-unused" foreColor="gray" default="unused" />
-    <fontcolor name="mod-deprecated" default="deprecated" />
-    <fontcolor name="mod-abstract" default="abstract" />
-    <fontcolor name="mod-static" default="static" />
-    <fontcolor name="mod-private" default="private" />
-    <fontcolor name="mod-package-private" default="private"/>
-    <fontcolor name="mod-protected" default="protected" />
-    <fontcolor name="mod-public" default="public" />
-    
-    <!--currently not used:-->
-<!--    <fontcolor name="mod-type-parameter-declaration" bgColor="lightGray"/>
-    <fontcolor name="mod-type-parameter-use" bgColor="green"/>
-    <fontcolor name="mod-undefined" foreColor="red"/>-->
-
-    <!--mark occurrences:-->
-    <fontcolor name="mark-occurrences" bgColor="ff38566f"/>
-
-    <!-- JavaStringTokenId categories -->
-    <fontcolor name="string-escape" default="string">
-        <font style="bold" />
-    </fontcolor>
-    <fontcolor name="string-escape-invalid" default="string">
-        <font style="italic" />
-    </fontcolor>
-    
-    <!-- JavaCharacterTokenId categories -->
-    <fontcolor name="character-escape" default="character">
-        <font style="bold" />
-    </fontcolor>
-    <fontcolor name="character-escape-invalid" default="character">
-        <font style="italic" />
-    </fontcolor>
-   
-    <!-- JavadocTokenId categories -->
-    <fontcolor name="javadoc-tag" default="comment">
-        <font style="bold" />
-    </fontcolor>
-    <fontcolor name="javadoc-identifier" default="comment"/>
-    <fontcolor name="javadoc-first-sentence" default="comment"/>
-    
-    <fontcolor name="html-tag" foreColor="428BDD" default="default"/>
-
-</fontscolors>
