--- praxis.live.editor/src/org/praxislive/ide/editor/resources/PLTheme-annotations.xml
@@ -1,100 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
-
-Copyright 1997-2010 Oracle and/or its affiliates. All rights reserved.
-
-Oracle and Java are registered trademarks of Oracle and/or its affiliates.
-Other names may be trademarks of their respective owners.
-
-
-The contents of this file are subject to the terms of either the GNU
-General Public License Version 2 only ("GPL") or the Common
-Development and Distribution License("CDDL") (collectively, the
-"License"). You may not use this file except in compliance with the
-License. You can obtain a copy of the License at
-http://www.netbeans.org/cddl-gplv2.html
-or nbbuild/licenses/CDDL-GPL-2-CP. See the License for the
-specific language governing permissions and limitations under the
-License.  When distributing the software, include this License Header
-Notice in each file and include the License file at
-nbbuild/licenses/CDDL-GPL-2-CP.  Oracle designates this
-particular file as subject to the "Classpath" exception as provided
-by Oracle in the GPL Version 2 section of the License file that
-accompanied this code. If applicable, add the following below the
-License Header, with the fields enclosed by brackets [] replaced by
-your own identifying information:
-"Portions Copyrighted [year] [name of copyright owner]"
-
-Contributor(s):
-
-The Original Software is NetBeans. The Initial Developer of the Original
-Software is Sun Microsystems, Inc. Portions Copyright 1997-2006 Sun
-Microsystems, Inc. All Rights Reserved.
-
-If you wish your version of this file to be governed by only the CDDL
-or only the GPL Version 2, indicate your decision by adding
-"[Contributor] elects to include this software in this distribution
-under the [CDDL or GPL Version 2] license." If you do not indicate a
-single choice of license, a recipient has the option to distribute
-your version of this file under either the CDDL, the GPL Version 2 or
-to extend the choice of license to its licensees as provided above.
-However, if you add GPL Version 2 code and therefore, elected the GPL
-Version 2 license, then the option applies only if the new code is
-made subject to such option by the copyright holder.
--->
-<!DOCTYPE fontscolors PUBLIC "-//NetBeans//DTD Editor Fonts and Colors settings 1.1//EN" "http://www.netbeans.org/dtds/EditorFontsColors-1_1.dtd">
-
-<fontscolors>
-    <fontcolor bgColor="611D1D" name="Breakpoint"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_BP"/>
-
-
-    <fontcolor bgColor="611D1D" name="Breakpoint"/>
-    <fontcolor bgColor="2D6035" name="CurrentPCLinePart"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_mixedBP"/>
-    <fontcolor bgColor="2D6035" name="OtherThreads_PC_BP"/>
-    <fontcolor bgColor="2D6035" name="CurrentPC2"/>
-    <fontcolor bgColor="2D6035" name="OtherThreads_BP_broken"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_BP_stroke"/>
-    <fontcolor bgColor="611D1D" name="debugger-mixed_BP"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_DCBP"/>
-    <fontcolor bgColor="2D6035" name="OtherThread_PC"/>
-    <fontcolor bgColor="2D6035" name="CurrentPC"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_multi_BPCBP"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_CBP"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_DBP_stroke"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_multi_DBPCBP"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_multi_BPCBP_broken"/>
-    <fontcolor bgColor="611D1D" name="OtherThreads_BP"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_BP_broken"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_mixedBP_broken"/>
-    <fontcolor bgColor="611D1D" name="debugger-multi_BPCBP"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_BP"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_DCBP_stroke"/>
-    <fontcolor bgColor="2D6035" name="OtherThread_BP_broken"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_CBP_stroke"/>
-    <fontcolor bgColor="2D6035" name="CurrentPC2LinePart"/>
-    <fontcolor bgColor="611D1D" name="OtherThread_BP"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_CBP_broken"/>
-    <fontcolor bgColor="2D6035" name="debugger-PC_DBP"/>
-    <fontcolor bgColor="611D1D" name="CondBreakpoint"/>
-    <fontcolor bgColor="611D1D" name="CurrentPC2_BP"/>
-    <fontcolor bgColor="2D6035" name="OtherThread_PC_BP"/>
-
-
-    <fontcolor bgColor="666666" name="OtherThreads_DBP"/>
-    <fontcolor bgColor="666666" name="DisabledBreakpoint"/>
-    <fontcolor bgColor="666666" name="debugger-multi_DBPCBP"/>
-    <fontcolor bgColor="666666" name="Breakpoint_broken"/>
-    <fontcolor bgColor="666666" name="DisabledBreakpoint_stroke"/>
-    <fontcolor bgColor="666666" name="CondBreakpoint_broken"/>
-    <fontcolor bgColor="666666" name="Breakpoint_stroke"/>
-    <fontcolor bgColor="666666" name="debugger-mixed_BP_broken"/>
-    <fontcolor bgColor="666666" name="CondBreakpoint_stroke"/>
-    <fontcolor bgColor="666666" name="DisabledCondBreakpoint"/>
-    <fontcolor bgColor="666666" name="debugger-multi_BPCBP_broken"/>
-    <fontcolor bgColor="666666" name="DisabledCondBreakpoint_stroke"/>
-    <fontcolor bgColor="666666" name="OtherThread_DBP"/>
-    <fontcolor bgColor="666666" name="CurrentPC2_DBP"/>
-</fontscolors>
