--- praxis.live.editor/src/org/praxislive/ide/editor/resources/highlights.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE fontscolors PUBLIC "-//NetBeans//DTD Editor Fonts and Colors settings 1.1//EN" "http://www.netbeans.org/dtds/EditorFontsColors-1_1.dtd">
-<fontscolors>
-    <fontcolor foreColor="ff666666" name="code-folding"/>
-    <fontcolor name="caret-color-insert-mode"/>
-    <fontcolor bgColor="ffe0e8f1" foreColor="black" name="block-search"/>
-    <fontcolor foreColor="ff4c5153" name="text-limit-line-color"/>
-    <fontcolor bgColor="ff204a87" name="nbeditor-bracesMatching-match-multichar"/>
-    <fontcolor bgColor="black" foreColor="ff555753" name="line-number"/>
-    <fontcolor name="hyperlinks" underline="blue"/>
-    <fontcolor bgColor="ff3d4341" name="highlight-caret-row"/>
-    <fontcolor foreColor="red" name="status-bar-bold"/>
-    <fontcolor bgColor="ff555753" name="selection"/>
-    <fontcolor name="status-bar"/>
-    <fontcolor bgColor="ffffb442" foreColor="black" name="inc-search"/>
-    <fontcolor foreColor="ffcc0099" name="synchronized-text-blocks-ext"/>
-    <fontcolor foreColor="ff666666" name="code-folding-bar"/>
-    <fontcolor bgColor="ffffb442" foreColor="black" name="highlight-search"/>
-    <fontcolor bgColor="ffef2929" name="nbeditor-bracesMatching-mismatch-multichar"/>
-    <fontcolor foreColor="ff996600" name="synchronized-text-blocks-ext-slave"/>
-    <fontcolor bgColor="ffdfdfdf" name="guarded"/>
-    <fontcolor bgColor="ff204a87" name="nbeditor-bracesMatching-match"/>
-    <fontcolor bgColor="ffef2929" name="nbeditor-bracesMatching-mismatch"/>
-</fontscolors>
