--- praxis.live.laf/build.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- You may freely edit this file. See harness/README in the NetBeans platform -->
-<!-- for some information on what you could do (e.g. targets to override). -->
-<!-- If you delete this file and reopen the project it will be recreated. -->
-<project name="org.praxislive.ide.laf" default="netbeans" basedir=".">
-    <description>Builds, tests, and runs the project org.praxislive.ide.laf.</description>
-    <import file="nbproject/build-impl.xml"/>
-</project>
