--- praxis.live.laf/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=32b57550
-build.xml.script.CRC32=19978fd4
-build.xml.stylesheet.CRC32=a56c6a5b@2.71.1
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=32b57550
-nbproject/build-impl.xml.script.CRC32=060d12ef
-nbproject/build-impl.xml.stylesheet.CRC32=238281d1@2.71.1
