--- praxis.live.laf/nbproject/project.properties
@@ -1,6 +1,0 @@
-javac.source=11
-javac.compilerargs=-Xlint -Xlint:-serial
-#license.file=../../../../../../GPL2CPE.txt
-#nbm.homepage=http://www.praxislive.org
-#nbm.module.author=Neil C Smith
-#nbm.needs.restart=true
