--- praxis.live.laf/src/org/praxislive/ide/laf/Bundle.properties
@@ -1,2 +1,0 @@
-OpenIDE-Module-Name=praxis.live.laf
-OpenIDE-Module-Short-Description=Praxis LIVE Look and Feel
